+++ app/Models/Score/Best/Model.php
@@ -9,8 +9,8 @@ use App\Libraries\ReplayFile;
 use App\Models\Beatmap;
 use App\Models\BeatmapModeStats;
 use App\Models\ReplayViewCount;
-use App\Models\Reportable;
 use App\Models\Score\Model as BaseModel;
+use App\Models\Traits;
 use App\Models\User;
 use App\Traits\WithDbCursorHelper;
 use Datadog;
@@ -23,7 +23,7 @@ use GuzzleHttp\Client;
  */
 abstract class Model extends BaseModel
 {
-    use Reportable, WithDbCursorHelper;
+    use Traits\Reportable, WithDbCursorHelper;
 
     public $position = null;
     public $weight = null;
