+++ app/Models/Traits/Reportable.php
@@ -3,8 +3,10 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace App\Models;
+namespace App\Models\Traits;
 
+use App\Models\User;
+use App\Models\UserReport;
 use App\Notifications\UserReportNotification;
 use PDOException;
